[Feat] 내구도 아이템 사용 구현
아이템 내구도에서 durabilityCost 빼기
내구도가 0이면 슬롯에서 아이템 삭제

+++ Assets/01 Scripts/UI/ItemSlotUI.cs
@@ -317,6 +317,6 @@ public class ItemSlotUI : MonoBehaviour,
 
     public void ChangeDurabilitySliderValue(int current, int max)
     {
-        _durabilitySlider.value = current / max;
+        _durabilitySlider.value = (float)current / (float)max;
     }
 }
